viewmodel -> bottomsheet-> fragment 로 itemClick 인자 전달하여 이동
