+++ app/src/main/java/kids/baba/mobile/presentation/viewmodel/EditMemberProfileBottomSheetViewModel.kt
@@ -26,6 +26,9 @@ class EditMemberProfileBottomSheetViewModel @Inject constructor(
     private val editProfileUseCase: EditProfileUseCase
 ) : ViewModel() {
 
+    var itemClick = {}
+    var dismiss: () -> Unit = {}
+
     private val _eventFlow = MutableEventFlow<EditMemberProfileEvent>()
     val eventFlow = _eventFlow.asEventFlow()
 
@@ -58,12 +61,16 @@ class EditMemberProfileBottomSheetViewModel @Inject constructor(
                     )) {
                         is Result.Success -> {
                             _eventFlow.emit(EditMemberProfileEvent.SuccessEditMemberProfile)
+                            itemClick()
+                            dismiss()
                         }
                         is Result.NetworkError -> {
                             _eventFlow.emit(EditMemberProfileEvent.ShowSnackBar(R.string.baba_network_failed))
+                            dismiss()
                         }
                         else -> {
                             _eventFlow.emit(EditMemberProfileEvent.ShowSnackBar(R.string.unknown_error_msg))
+                            dismiss()
                         }
                     }
                 }
